Fix issue #34: Face ID support.
Face ID is supported out of the box, but I updated the API to allow developers to determine the biometric type used for auth (currently face or touch).

This was done by adding a property to the isAvailable method allowing it to return a string with 'face' or 'touch' type.